pyproject.toml
[tool.poetry]
name = "nix-airflow-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
# https://airflow.apache.org/docs/stable/installation.html#extra-packages
apache-airflow = { version ="^1.10.10", extras = ["postgres", "kubernetes", "celery"]}
typing_extensions = "^3.7.4"
typeit = "^0.27.1"

[tool.poetry.dev-dependencies]
mypy = "^0.770"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
